+++ tasmota/language/zh_TW.h
@@ -836,8 +836,8 @@
 #define D_SENSOR_HALLEFFECT    "HallEffect"
 #define D_SENSOR_EPD_DATA      "EPD Data"
 #define D_SENSOR_MCP2515_CS    "MCP2515 CS"
-#define D_SENSOR_HRG15_RX      "HRG15 RX"
-#define D_SENSOR_HRG15_TX      "HRG15 TX"
+#define D_SENSOR_HRG15_RX      "HRG15 Rx"
+#define D_SENSOR_HRG15_TX      "HRG15 Tx"
 
 // Units
 #define D_UNIT_AMPERE "安培"