+++ tasmota/tasmota_template.h
@@ -170,7 +170,7 @@ enum UserSelectablePins {
   GPIO_I2S_IN_DATA,  GPIO_I2S_IN_CLK,  GPIO_I2S_IN_SLCT,
   GPIO_INTERRUPT,
   GPIO_MCP2515_CS,                     // MCP2515 Chip Select
-  GPIO_HRG15_RX, GPIO_HRG15_TX,        // Hydreon RG-15 rain sensor serial interface
+  GPIO_HRG15_TX, GPIO_HRG15_RX,        // Hydreon RG-15 rain sensor serial interface
   GPIO_SENSOR_END };
 
 enum ProgramSelectablePins {
@@ -361,7 +361,7 @@ const char kSensorNames[] PROGMEM =
   D_SENSOR_I2S_IN_DATA  "|" D_SENSOR_I2S_IN_CLK  "|" D_SENSOR_I2S_IN_SLCT  "|"
   D_SENSOR_INTERRUPT "|"
   D_SENSOR_MCP2515_CS "|"
-  D_SENSOR_HRG15_RX "|" D_SENSOR_HRG15_TX
+  D_SENSOR_HRG15_TX "|" D_SENSOR_HRG15_RX
   ;
 
 const char kSensorNamesFixed[] PROGMEM =
@@ -794,8 +794,8 @@ const uint16_t kGpioNiceList[] PROGMEM = {
   AGPIO(GPIO_AS608_RX),
 #endif
 #ifdef USE_HRG15
-  AGPIO(GPIO_HRG15_RX),
   AGPIO(GPIO_HRG15_TX),
+  AGPIO(GPIO_HRG15_RX),
 #endif
 
 /*-------------------------------------------------------------------------------------------*\