+++ CHANGELOG.md
@@ -9,6 +9,7 @@ All notable changes to this project will be documented in this file.
 - Support for IEM3155 Wattmeter (#12940)
 - Berry support for vararg
 - Command ``Subscribe2 ...`` to subscribe to a MQTT topic without appended "/#" (#12858)
+- Support for Hydreon RG-15 Solid State Rain sensor (#12974)
 
 ### Changed
 - Shelly EM template needs to use GPIO ADE7953_IRQ_2